build.sh: use show_info and show_err
In fact, print_info and print_err in the build.sh script
are defined in common_functions.sh. However,if an unexpected event
occurs during compilation, the build.sh script may not be able to
call these two functions, so show_info and show_err are defined
separately in build.sh.

Signed-off-by: carbon <carbon@milkv.io>